Обновлены файлы проекта

+++ src/main/java/com/example/bdhotel/models/Rooms.java
@@ -19,7 +19,7 @@ public class Rooms {
     private String className;
 
     @Column(name = "costPerDay")
-    private double costPerDay;
+    private int costPerDay;
 
 
     @Column(name = "status")
